Module Tezos_stdlib__Circular_buffer

This module implements a bufferisation abstraction to store temporary raw data chunks (as bytes) when chunks are read sequentially. The function write allows to store chunks in the buffer and the function read to read them from the buffer.

The global contract is that if we write consecutively d1;d2 onto the buffer. Then we have to fully read d1 and d2, in that order.

This contract is not enforced by the library, it is the user responsability to respect it.

If the circular buffer is full, a new temporary buffer is allocated to store the chunk of data to be written.