Module Tezos_stdlib__Tag

Tags and tag sets. Tags are basically similar to a plain extensible variant type, but wrapped with metadata that enables them to be printed generically and combined into tag sets where each tag is either not present or associated with a specific value.

They are primarily intended for use with the `Logging` module but it would probably be reasonable to use them for other purposes.